requirements.txt
aniso8601==4.0.1
asn1crypto==0.24.0
backcall==0.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
copper==0.9.0
cryptography==2.4.1
cycler==0.10.0
decorator==4.3.0
Flask==1.0.2
Flask-RESTful==0.3.6
idna==2.7
ipython==7.2.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.1
mkl-fft==1.0.6
mkl-random==1.0.1
numpy==1.15.4
olefile==0.46
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycparser==2.19
pyFirmata==1.0.3
Pygments==2.3.1
pylibftdi==0.17.0
pyOpenSSL==18.0.0
pyparsing==2.3.0
pyserial==3.4
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.7
pyudev==0.21.0
pyusb==1.0.2
pyzmq==17.1.2
requests==2.20.1
scikit-learn==0.20.1
scipy==1.1.0
six==1.11.0
some-package==0.1
torch==1.0.0
torchfile==0.1.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
urllib3==1.23
usbinfo==1.1.1
visdom==0.1.8.5
wcwidth==0.1.7
websocket-client==0.54.0
Werkzeug==0.14.1
